Move cursor class, add factory
